vh: sched: Protect uclamp_update_active() with uclamp_is_used()
We should not do uclamp operation while it's not being used.

No special problem was observed - but we can potentially end up calling
uclamp_rq_inc/dec_id() before the key is enabled and miss up the state
of uclamp in the scheduler.

Better be safe and add the protection to avoid potential problems.

Bug: 261695814
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I1dc7a5b3c6ad064997fd87bc77c4d073cb79175e