eh: Fix compression thread deadlock due to nr_request becoming negative
The "eh: Make the compression thread more efficient" patch changed the
order in which nr_request is incremented with respect to compression
kickoff. As a result, a narrow window for a race condition opened up which
made it possible to deadlock the compression thread.

Because compression is kicked off before nr_request is incremented, the
compression thread may finish processing a new compression request before
nr_request is incremented for that request. This would then lead to the
compression thread looping again on the nr_request != 0 check, and then
deadlock forever waiting for a compression completion interrupt that'll
never arrive because there are no compression requests in flight.

Fix it by restoring the original order in which nr_request is incremented
with respect to compression kickoff. There's no actual need for nr_request
to be incremented after compression kickoff, so this is safe.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>